Initial commit of the observability code.
Bazel builds are not present because they do not work. (I will upload these in a separate branch.)

Container builds are not yet there, but will be soon.

The step after that will be to simply upload these containers to a container registry in Github, using a Github pipeline runner.