Fix firewall rule management.
* Fix typo in proposal to remove firewall rules.
* Fix signature of remove command being called.
* Batch multiple operations so we can at least remove / modify / add
  multiple rules in one proposal.

Before we do the batching, we sort the rules by modification slot
then we reverse the order of the rules list, and then we do the
batching.  This ensures that the modifications happen in an order
that will not conflict between each other (what does it mean "add
rule in position 3" if rule in position 2 was removed prior to
that?).