Introduce interrupt and retry handling to initial definition sync.
We have been quite good with regards to handling interrupts / SIGTERMs almost from the very start in MSD.  However, a small oversight with regards to initial sync caused https://github.com/dfinity/dre/issues/314 , which in turn causes the daemon to freeze upon shutdown if any definition failed its initial sync.  The root cause is that the thread terminates on its own, not polling its end of the ender channel used to shut it down.

This PR fixes the issue.

In addition to that, we now also retry initial syncs just as failed incremental syncs are retried.  This adds robustness in the face of temporary network outages just as MSD is starting to execute.