feat(dre): Add WhatifAnalysis command for subnet analysis
This commit adds a new command, WhatifAnalysis, to the Subnet module. The WhatifAnalysis command allows users to perform subnet analysis by simulating node additions and removals. It takes in parameters such as the nodes to add and remove, subnet ID, and initial nodes. The command executes the analysis by calling the decentralization_change method in the Runner module.

This allows community members to explore and experiment, to see how particular subnet topology changes would affect the subnet decentralization.

Parts of implementation:
* [x] Add a WhatIf subcommand: `dre subnet whatif <...>`
* [x] Adjust the decentralization_change function to accept (optional) arbitrary initial node list in the subnet
* [ ] Add documentation