Refactor authentication detection.
Improve error messages shown to user when authentication has
problems rather than bombing out with a cryptic "invalid auth"
message.

Peel off from the destructuring in the expression tree the code
path where the private key file is handled.  This makes it clear
to the maintainers of this code that the two paths are mutually
exclusive rather than there being an acceptable mix of private
key and HSM arguments.