Restore original, intended behavior of cargo deny.
--warn unmaintained causes Cargo Deny to ignore the exceptions we
have coded into deny.toml.

Therefore removed.

Also restored -D to intentionally cause failures when this is
executed, so that a commit is not possible to be merged when
it introduces a problem.

Also only run this when it's a pull request.  No need to run
this on merge (it already ran before) or on tag (it already
ran before).