feat(dre): Provide an estimate of node types in the `dre registry`
The node type information is not yet associated with node records in the registry.
There is an ongoing effort to rectify that, and to achieve that we need an accurate estimate of the node types based on the information available in the node operator record: rewardable_nodes.

What the code does is:
- iterate over all nodes, and
- for each node fetch the node operator
- count the number of rewardable nodes of in each type (implementation detail: this is done by subtracting the rewardable nodes count)
- estimate the node type based on the next available rewardable_nodes entry (not yet assigned to some other node record)

Again, this is only an estimate, and the correct information will be added to the registry in the future by the NNS team. At that point, we'll want to adjust the code.