feat: Add functionality to ensure that at least one node from each operator is in use
If a node operator does not have any node assigned to a subnet, we cannot know if these nodes are in good shape. When there is no reduction in decentralization (!), let's try to have at least one node of each operator in a subnet, since we will have better insights into node health.
The new functionality can be called with:
```
dre network --ensure-operator-nodes-assigned
```

Also, a few minor changes in the PR:
- silence ic-admin notices since we are running it from automation
- Refactor decentralization change handling, to make impact of a change on decentralization more consistent across operations
- Rename the "heal" subcommand with "network --heal", but keep backwards compatibility. We may remove it at some point in the future