Allow DRE vote and other operations to repair the PKCS#11 context that might have gone bad.
When suspending a Linux system, or when accidentally touching a loose NitroKey, the PKCS#11 context used to talk to it gets broken and does not know how to recover.  When we attempt to sign messages, we now detect this situation, drop the prior context, and create a new one.  If this succeeds, all good.  If this fails, we report the original error to the user, and notify them as well that we could not regenerate a valid PKCS#11 context.