feat: update cordoned features handling and support for local file
Previous implementation was using the local file for cordoned features *only if* --offline was provided. That means that the entire dre tool had to be forced to go offline.
With this change, one can simply provide `--cordoned-features-file <file>` on the CLI, and the local file will override the github version