feat(dre): add ability to omit/ignore a list of nodes when healing the entire network
The list of omitted nodes is used internally when healing and optimizing the network since certain steps may add nodes to subnets and we don't want to use the same nodes in subsequent steps. So a kind of a node reservation.
It may make sense to expose this list to the command line user as well,  e.g. in case one does not want particular nodes to be considered healthy and to be added to subnets, even though nothing is obviously wrong with the nodes. This way the list of omited nodes does not start empty.

This makes the omit_nodes work the same as omit_subnets, that was introduced earlier.

Also, several minor refactorings and bug fixes.