Remove all mentions of the forum post link in the summary of election proposals.
Rationale: the forum post link is already in the proposal.  Requiring a forum post to compose the proposal summary creates a chicken-and-egg problem whereby the proposal requires the forum post and the forum post requires the proposal.  Hacks in the codebase to get around that problem include, but are not limited to, generating the summary twice (once for test run and once for final run).

Removal of this makes the codebase cleaner and makes it harder to make a programming mistake that would result in no forum post, or the wrong forum post link, or a forum post without a proposal being submitted.