Major refactor of proposal structures and processing of proposals.
This is being submitted for review to gather questions and observations from you,
because I am positive a refactor of this kind will raise many questions and I
want to address them all before merging this.

Changes of note:

* `dre vote` no longer votes automatically by default, for consistency.
  Pass `--yes` to autovote.
* This PR opens the door to a new kind of proposal (not based on IC admin)
  that will use the governance canister to directly submit proposals.
  That will allow us to delete some more code.  Currently we only have the
  `ProposableViaIcAdmin` trait, but soon this will expand to just be a
  Proposable trait that will encompass both proposals sent thru IC admin
  and also thru governance canister.
* The `forum_post_link` struct member has finally been deleted.  All forum
  post handling happens right before forrealz submission of a proposal,
  instead of the sui generis handling that took place before.

I regret that the full code deletion cannot be completed at this time,
but this refactor only added 50 lines of code, with the benefit of unified
handling of many user-exposed options.