Refactor: consolidate forum parameters and confirmation parameters into a single structure.
The 'SubmissionParameters' structure has sprung to existence to include the `ForumParameters`
and `ConfirmationModeOptions' structs, and is now used when appropriate (in operations that
require forum post parameters and can be affected by confirmation modes, e.g., dry run,
confirm, unconditional).

These changes, together with prior changes that consolidate forum post management and proposal
simulation / execution, make the code more maintainable, more explicit about what is taking
place, and pave the way for imbuing the `Proposable` types with knowledge about how to deal
with the forum based on what the type of `Proposable`.

Noteworthy in light of the last sentence: now the Submitter type method `propose()` accepts
a `ProposalExecution` (tied to a `Proposable`) next to a forum post type.  Soon, the type
of forum post can be dictated by the `ProposalExecution` itself, which makes sense, as we
already have fixed forum post types for different proposals, but this knowledge is currently
(incorrectly) hardcoded in an ad-hoc manner throughout various parts of the code.