Use Bazel to build DRE CLI when running release controller through Bazel
Our release controller can be run with `bazel run`.  When this takes place,
currently and by default the release controller will attempt to use whatever
`dre` binary is in the user's `$PATH`.

This obviously fails when there's no such binary.

To address this, and to ensure that the release controller is running with
an up-to-date version of `dre`, this change makes Bazel build the `dre`
binary before starting the release controller, and then makes the release
controller use the Bazel-built `dre` program, but only when running under
Bazel.

This commit introduces support for an optional `cli_path` parameter in the
`DRECli` class constructor, allowing users to specify a custom path to the
`dre` CLI executable. This change is also useful when running the code
within different environments or when testing with CLI mocks.

The following files were updated:
- `release-controller/dre_cli.py`
- `release-controller/dryrun.py`
- `release-controller/reconciler.py`