This is a twofer: fix ugly error message in annotator, and enhance docs.
The first thing in this commit is to remove this message from the standard error:

```
error: no note found for object 7d4e057cbb17f4bf2a108d547402cb5e7bf3eacc.
```

which is unnnecessary.  I even added the functionality to the `notes()` function before, but neglected to call it when it was appropriate.

The second thing is the URL to the Jira ops board which should exclude closed alerts because they are noise and should not be shown by default.

I'll handle rollout of the annotator after this.