Speed up target determination.
Prior to this PR, we were executing the bazel cquery and the target determination as sequential operations, each of which
checked out a commit -- so every commit was checked out twice.

We now check out the commit only once, and make the target determination parallel with the bazel cquery, using the same
local checkout in both cases.

This provides a substantial speedup primarily due to a reduction in disk access waste, but also due to the fact that now
more cores can be active churning results.