This pull request addressess https://dfinity.atlassian.net/browse/DRE-395 .
The main feature added is support for HostOS in the commit annotator and
the reconciler.

* Commit annotations for HostOS.
* Forum posts and release notes composed for HostOS.
* Submissions of proposals for HostOS election.

HostOS is only published in its `base` version.  Feature releases do not
get elected for HostOS.

In addition to that, reconciler gains a feature (which will have to be
properly configured in the Kubernetes pod) whereby commit annotations are
fetched from the annotator instead of from the local Git repository. With
this change, the reconciler asks the annotator for commit information
from its own copy of the repository, where annotations are always more
up to date than whatever is on the main repository; reconciler thus no
longer needs to fetch annotations on every cycle, enormously reducing
the amount of work reconciler needs to do.  If an annotation is missing
during the preparation of the release notes stage, reconciler will log
that as a warning, and move on to the next release it needs to prepare.

The annotator now has an HTTP API for the purpose of serving annotations
to the reconciler.  This API can be disabled using flag value
`--api-port=0` and is also not active when the `--loop-every=0` flag is
specified (for a single run rather than running in a loop).

The feature mentioned above can be disabled when interactively running the
reconciler by passing the `--commit-annotator-url=local` flag to the
`bazel run //release-controller:reconciler -- ...` command.  Under this
mode of operation, the reconciler behaves exactly as it behaved before,
with regards to annotations.

The reconciler now keeps track of the state of each release/version/OS
combo, and uses that state (in-memory) to decide whether to retry or repeat
an operation.  This saves enormously (in particular) on queries performed
to Discourse, which in the past has resulted in Discourse throttling the
reconciler and causing it to malfunction.  In addition, thanks to this
new state tracking feature, we can keep track of the state of each phase
via telemetry, since this information is now published as Prometheus
metrics.

The release notes generator can now be run interactively on the command
line.  Run `bazel run //release-controller/release-notes -- --help` for
more information on what parameters it needs.  The obsolete release notes
generator for HostOS has been deleted -- you can use the established
release notes generator for that.  Release notes generation for GuestOS
now only contemplates changes affecting the `update.img.zst` file that
gets published as the official updater for GuestOS -- changes that affect
SetupOS now fall under the HostOS release notes generation process.

The `GitAnnotator` class is no longer available as a context manager.  Now
the caller must construct it by passing a `GitRepo` object.  This change,
along with other related changes, speeds up access and modification of
git notes, and therefore improves operation of the annotator.

The initial `git fetch` done by the `GitRepo` class no longer prunes
remote refs.  Pruning remote refs destroys the incremental progress that
the annotator has made prior to pushing the annotations to the remote
server, and there is no other use case that requires pruning of the
local refs.

The `GovernanceCanister` client used (as of the last commit) to retrieve
submitted proposals has been deleted.  In its stead, the `dre` CLI tool
we were using before for other governance-related matters is now used.
This tool is also used, as of this commit, to submit proposals for HostOS
elections.

The `ICPrometheus` client formerly used only for retrieving the active
GuestOS versions can now retrieve the active HostOS versions too.

The reconciler has been instructed to ignore all releases before this
week's commit, to avoid touching older forum posts, which might be
modified due to the change in how we order forum posts when publishing
them.  The new order is as such:

* Top post of the thread: standard GuestOS release info.
* Second post of the thread: HostOS release info.
* From then on, feature releases of GuestOS.

The commit annotator now pulls the leg of the watchdog on each successful
annotation, preventing it from crashing when there is a long string of
commits that need to be annotated.  The watchdog continues to crash the
commit annotator (as designed) if annotation gets stuck in any other way.

The dependency on the `webbrowser` module has been deleted.

The dependency footprint of the commit annotator has been reduced.