Address crashlooping of the commit annotator.
There was one commit in the history of the IC repo that causes a perennial crash loop when querying for whether the commit belongs in the HostOS changelog.

To address this issue, the change allows for commits to be annotated as `Failed` when the target determinator fails consistently on that commit.

In addition to that:

* This commit changes the commit annotation function to clean up the working directory prior to attempting annotation, which accelerates execution of the target determinator as it no longer needs to autocreate a working directory separately.
* This commit also imposes a 500 commit limit to how many commits it will annotate per branch, to avoid going back 3400 commits (which would take forever).  The count of commits is determined experimentally by counting commits between one month worth of releases, and should be sufficient.