Wrap the entire phase of release notes preparation into the phase decorator.
With this, we will get telemetry of successful release notes preparation even when the notes do not need to be prepared, enhancing the consistency of the telemetry we get.

This also allows us to capture and record telemetry for any exceptions within the check portion of the phase, which was previously not wrapped in the phase decorator.