Improve  HostOS release notes generation.
In this commit, we change the query we use for Bazel and `target-determinator`
to calculate which targets have been affected on each (main) commit.

This change is powered by the newly-created ability of the `release-notes`
program (invokable via Bazel) to use a new implementation of change determination
based on a 100% cache-free local `internal` calculation of commit changes,
whereby instead of consulting Git notes, the release-notes program manually
recomputes everything for every single commit.

We also add the targets that target-determinator discovered as changed
in the specific Git note saved for the purpose.

Finally, a small change to the `progressbar` code allows concomitant messages
to print cleanly without tampering with the progress bar.  This does not affect
the operation of the program in production.