feat: add user and root home directories for containers
This commit adds support for creating a user home directory in the OCI image.

It includes a `home` rule with two entries for `/root` and `/home/user`, providing appropriate UID, GID, and shell configurations for each user.

This fixes the release controller crashing because it wants to write to `/.cache`.