fix(reconciler): correct issue with reconciler and Slack messages.
When we moved to Ubuntu 24.04 for the container package of the reconciler, something changed
in the way that the container was built, such that Bazel-wrapped Python binaries (such as the
reconciler) silently failed to load the system certificate store, leading to Python
`urllib.request.urlopen` failing to load HTTPS resources.  Since `urllib` is how the Slack
SDK posts announcements about GuestOS and HostOS, that begun failing (fortunately not in
a way that prevented making progress with the release process).

This failure does not affect usage of the `requests` because `requests` uses the `certifi`
certificate store by default (although if we were using a private certificate authority for
some of our requested resources, this would not work and it would require the use of the
environment variable `REQUESTS_CA_BUNDLE` to override that).

This failure, strangely, also does not affect plain `urllib.request.urlopen` from the
container's Python.  When that is invoked manually within the container, it works OK.
It only seems to affect any script wrapped by Bazel and packaged up in the container.

The fix here is to pay attention to which files the `urllib` (and `ssl`) module attempt to
read when creating an SSL context, and symlinking one of those files to the system certificate
store in `/etc/ssl/certs`.  Which is exactly what this fix does.