Disable team reviews for changelogs of HostOS and GuestOS
With this, Slack announcements will continue to happen when release notes are prepared for review.  However,
no teams will be pinged on Slack or listed in the document for cross-out to mark review as reviewed.
This in turn means that the reconciler will proceed immediately to publishing the changelog as-is to the
Github repository, in preparation for creation of the proposal.

Motivated by teams asking not to be reviewing the changelog anymore.

Closes https://dfinity.atlassian.net/browse/DRE-465 .