feat(release-controller): log post differences before updating
When a post is detected to differ from the expected content, log the differences before updating the post.

This is going to help us address a situation whereby a log line appears to indicate that posts are updated
after a restart of the release controller, but no visible content changes appear on Discourse.  It will
also be useful to diagnose and revert the effects of a hypothetical a bad prod run whereby old posts are
updated in unexpected ways.