Stop release controller from updating all Discourse posts on its first after-boot cycle.
Analysis of what release controller was doing (thanks to the post diff feature added yesterday)
led us to discover that posts were generated with a carriage return at the end, which Discourse
happily accepts but then promptly chomps.  Thus, every time the code tried to update the posts,
it would conclude that the Discourse-served data and the local post data were different, and
promptly re-update the posts.  This does not happen on subsequent reconciliations after the
first one, because there is logic that keeps (in-memory) whether a post has already been updated
or not, and does not attempt to update the posts after that (saving Discourse API requests as
the posts won't be checked more than once during the lifecycle of release controller).

We solve this problem by chomping the trailing whitespace out of the post generator that
creates the posts for Discourse.