Fix(release-controller): annihilate two bugs in changelog processing
The first bug caused the release controller not to be able to understand the Markdown
generated from Google Docs (underlines with equals for titles instead of hashes as
prefix for titles).

The second bug involves the HostOS changelog being posted in lieu of the GuestOS
changelog.

This PR fixes both.