Make forum post updates non-fatal errors.
Prior to this PR, we were handling a fatal forum post update error within the ForumClient, which meant the top level code did not get a chance to retry posting the forum post again later.  As a consequence, forum posts remained not updated, and manual intervention was required.

This behavior was added because the situation *before* that behavior was that a release was never finished processing whenever Discourse had a migraine.

This PR proposes a putatively superior behavior: handle the error by making it non-fatal directly in the loop of the reconciler, such that the reconciler will still page us when Discourse is having problems (we mark the release as failed in 'forum post update' phase), but because the update never actually "happened" from the standpoint of the reconciler loop, it still gets retried the next loop, eventually succeeding.

Submitted for your consideration.